Update pan-tilt, allowing for angle offsets, and slowly stepping towards goal angle

+++ prototype/format/src/device_to_server.rs
@@ -17,8 +17,8 @@ pub enum DeviceToServer {
 #[derive(Serialize, Deserialize, Debug, Clone, Copy)]
 #[cfg_attr(feature = "defmt", derive(Format))]
 pub struct PanTiltStatus {
-    pub pan_deg: f32,
-    pub tilt_deg: f32,
+    pub pan_deg: i32,
+    pub tilt_deg: i32,
 }
 
 pub type MicArraySample = [i16; 4];
